+++ test/a11y/lint/rules/robust/test_link_to_missing_accessible_name_erb.rb
@@ -5,7 +5,7 @@ require "test_helper"
 module A11y
   module Lint
     module Rules
-      class TestMissingAccessibleNameErb < Minitest::Test
+      class TestLinkToMissingAccessibleNameErb < Minitest::Test
         # <%= link_to("", "/path", class: "icon") %>
         def test_link_to_with_empty_text_reports_offense
           offenses = run_fixture("link_to_empty_text")
@@ -17,7 +17,7 @@ module A11y
           )
           assert_equal(1, offenses[0].line)
           assert_equal(
-            "MissingAccessibleName",
+            "LinkToMissingAccessibleName",
             offenses[0].rule
           )
         end
@@ -163,102 +163,6 @@ module A11y
           assert_equal(1, offenses.length)
         end
 
-        # <%= button_tag("", class: "icon") %>
-        def test_button_tag_with_empty_text_reports_offense
-          offenses = run_fixture("button_tag_empty_text")
-
-          assert_equal(1, offenses.length)
-          assert_equal(
-            offense_message("button_tag"),
-            offenses[0].message
-          )
-          assert_equal(
-            "MissingAccessibleName",
-            offenses[0].rule
-          )
-        end
-
-        # <%= button_tag("Submit") %>
-        def test_button_tag_with_text_passes
-          offenses = run_fixture("button_tag_with_text")
-
-          assert_empty(offenses)
-        end
-
-        # <%= button_tag("", class: "icon", aria: { label: "Submit" }) %>
-        def test_button_tag_with_aria_hash_label_passes
-          offenses = run_fixture("button_tag_aria_hash_label")
-
-          assert_empty(offenses)
-        end
-
-        # <%= button_tag("", class: "icon", "aria-label" => "Submit") %>
-        def test_button_tag_with_string_aria_label_passes
-          offenses = run_fixture("button_tag_string_aria_label")
-
-          assert_empty(offenses)
-        end
-
-        # <%= button_tag(class: "button-icon") do %>
-        #   <%= inline_svg("icon.svg") %>
-        # <% end %>
-        def test_button_tag_with_block_reports_offense
-          offenses = run_fixture("button_tag_block")
-
-          assert_equal(1, offenses.length)
-          assert_equal(
-            offense_message("button_tag"),
-            offenses[0].message
-          )
-        end
-
-        # <%= button_tag(class: "button-icon", aria: { label: "Menu" }) do %>
-        #   <%= inline_svg("icon.svg") %>
-        # <% end %>
-        def test_button_tag_with_block_and_aria_label_passes
-          offenses = run_fixture("button_tag_block_aria_label")
-
-          assert_empty(offenses)
-        end
-
-        # <%= button_tag class: "button-icon" do %>
-        #   <%= inline_svg("icon.svg") %>
-        # <% end %>
-        def test_button_tag_with_block_without_parens_reports_offense
-          offenses = run_fixture("button_tag_block_no_parens")
-
-          assert_equal(1, offenses.length)
-        end
-
-        # <%= button_tag class: "button-icon", aria: { label: "Menu" } do %>
-        #   <%= inline_svg("icon.svg") %>
-        # <% end %>
-        def test_button_tag_with_block_without_parens_and_aria_label_passes
-          offenses = run_fixture("button_tag_block_no_parens_aria_label")
-
-          assert_empty(offenses)
-        end
-
-        # <%= button_tag(class: "button-icon",
-        #                type: "button") do %>
-        #   <%= inline_svg("icon.svg") %>
-        # <% end %>
-        def test_multiline_button_tag_with_block_reports_offense
-          offenses = run_fixture("button_tag_multiline_block")
-
-          assert_equal(1, offenses.length)
-        end
-
-        # <%= button_tag(class: "button-icon",
-        #                aria: { label: "Menu" }) do %>
-        #   <%= inline_svg("icon.svg") %>
-        # <% end %>
-        def test_multiline_button_tag_with_block_and_aria_label_passes
-          offenses = run_fixture("button_tag_multiline_block_aria_label")
-
-          assert_empty(offenses)
-        end
-
         # <%= link_to("", "/path") %>
         def test_sets_filename_on_offense
           offenses = run_fixture(
@@ -270,59 +174,6 @@ module A11y
             "app/views/index.html.erb",
             offenses[0].filename
           )
-        end
-
-        # <%= button_tag(class: "button-icon") do %>
-        #   <%= inline_svg("icon.svg") %>
-        #   <%= t(".label") %>
-        # <% end %>
-        def test_button_tag_with_block_and_text_passes
-          source = <<~ERB
-            <%= button_tag(class: "button-icon") do %>
-              <%= inline_svg("icon.svg") %>
-              <%= t(".label") %>
-            <% end %>
-          ERB
-
-          offenses = run_source(source)
-
-          assert_empty(offenses)
-        end
-
-        # <%= button_tag class: "button-icon" do %>
-        #   <%= inline_svg("icon.svg") %>
-        #   <%= t(".label") %>
-        # <% end %>
-        def test_button_tag_with_block_without_parens_and_text_passes
-          source = <<~ERB
-            <%= button_tag class: "button-icon" do %>
-              <%= inline_svg("icon.svg") %>
-              <%= t(".label") %>
-            <% end %>
-          ERB
-
-          offenses = run_source(source)
-
-          assert_empty(offenses)
-        end
-
-        # <%= button_tag(class: "button-icon",
-        #                type: "button") do %>
-        #   <%= inline_svg("icon.svg") %>
-        #   <%= t(".label") %>
-        # <% end %>
-        def test_multiline_button_tag_with_block_and_text_passes
-          source = <<~ERB
-            <%= button_tag(class: "button-icon",
-                           type: "button") do %>
-              <%= inline_svg("icon.svg") %>
-              <%= t(".label") %>
-            <% end %>
-          ERB
-
-          offenses = run_source(source)
-
-          assert_empty(offenses)
         end
 
         # <%= link_to("#", class: "icon") do %>
@@ -387,13 +238,13 @@ module A11y
 
         def run_fixture(name, filename: "test.html.erb")
           source = file_fixture(
-            "missing_accessible_name/erb/#{name}.html.erb"
+            "link_to_missing_accessible_name/erb/#{name}.html.erb"
           )
-          ErbRunner.new([MissingAccessibleName]).run(source, filename:)
+          ErbRunner.new([LinkToMissingAccessibleName]).run(source, filename:)
         end
 
         def run_source(source, filename: "test.html.erb")
-          ErbRunner.new([MissingAccessibleName]).run(source, filename:)
+          ErbRunner.new([LinkToMissingAccessibleName]).run(source, filename:)
         end
       end
     end