+++ test/a11y/lint/rules/robust/test_link_to_missing_accessible_name_phlex.rb
@@ -5,7 +5,7 @@ require "test_helper"
 module A11y
   module Lint
     module Rules
-      class TestMissingAccessibleNamePhlex < Minitest::Test
+      class TestLinkToMissingAccessibleNamePhlex < Minitest::Test
         # link_to("", "/path", class: "icon")
         def test_link_to_with_empty_text_reports_offense
           source = <<~RUBY
@@ -21,7 +21,7 @@ module A11y
           assert_equal(1, offenses.length)
           assert_equal(offense_message("link_to"), offenses[0].message)
           assert_equal(3, offenses[0].line)
-          assert_equal("MissingAccessibleName", offenses[0].rule)
+          assert_equal("LinkToMissingAccessibleName", offenses[0].rule)
         end
 
         # external_link_to("", "https://example.com", class: "icon")
@@ -179,107 +179,6 @@ module A11y
           assert_equal(1, offenses.length)
         end
 
-        # button_tag("", class: "icon")
-        def test_button_tag_with_empty_text_reports_offense
-          source = <<~RUBY
-            class TestView < Phlex::HTML
-              def view_template
-                button_tag("", class: "icon")
-              end
-            end
-          RUBY
-
-          offenses = run_linter(source)
-
-          assert_equal(1, offenses.length)
-          assert_equal(offense_message("button_tag"), offenses[0].message)
-          assert_equal("MissingAccessibleName", offenses[0].rule)
-        end
-
-        # button_tag("Submit")
-        def test_button_tag_with_text_passes
-          source = <<~RUBY
-            class TestView < Phlex::HTML
-              def view_template
-                button_tag("Submit")
-              end
-            end
-          RUBY
-
-          offenses = run_linter(source)
-
-          assert_empty(offenses)
-        end
-
-        # button_tag("", class: "icon", aria: { label: "Submit" })
-        def test_button_tag_with_aria_hash_label_passes
-          source = <<~RUBY
-            class TestView < Phlex::HTML
-              def view_template
-                button_tag("", class: "icon", aria: { label: "Submit" })
-              end
-            end
-          RUBY
-
-          offenses = run_linter(source)
-
-          assert_empty(offenses)
-        end
-
-        # button_tag("", class: "icon", "aria-label" => "Submit")
-        def test_button_tag_with_string_aria_label_passes
-          source = <<~RUBY
-            class TestView < Phlex::HTML
-              def view_template
-                button_tag("", class: "icon", "aria-label" => "Submit")
-              end
-            end
-          RUBY
-
-          offenses = run_linter(source)
-
-          assert_empty(offenses)
-        end
-
-        # button_tag(class: "button-icon") do
-        #   span(class: "icon-menu")
-        # end
-        def test_button_tag_with_block_reports_offense
-          source = <<~RUBY
-            class TestView < Phlex::HTML
-              def view_template
-                button_tag(class: "button-icon") do
-                  span(class: "icon-menu")
-                end
-              end
-            end
-          RUBY
-
-          offenses = run_linter(source)
-
-          assert_equal(1, offenses.length)
-          assert_equal(offense_message("button_tag"), offenses[0].message)
-        end
-
-        # button_tag(class: "button-icon", aria: { label: "Menu" }) do
-        #   span(class: "icon-menu")
-        # end
-        def test_button_tag_with_block_and_aria_label_passes
-          source = <<~RUBY
-            class TestView < Phlex::HTML
-              def view_template
-                button_tag(class: "button-icon", aria: { label: "Menu" }) do
-                  span(class: "icon-menu")
-                end
-              end
-            end
-          RUBY
-
-          offenses = run_linter(source)
-
-          assert_empty(offenses)
-        end
-
         # link_to("#", class: "icon") do
         #   inline_svg("icon.svg")
         #   span { t(".label") }
@@ -289,27 +188,6 @@ module A11y
             class TestView < Phlex::HTML
               def view_template
                 link_to("#", class: "icon") do
-                  inline_svg("icon.svg")
-                  span { t(".label") }
-                end
-              end
-            end
-          RUBY
-
-          offenses = run_linter(source)
-
-          assert_empty(offenses)
-        end
-
-        # button_tag(class: "button-icon") do
-        #   inline_svg("icon.svg")
-        #   span { t(".label") }
-        # end
-        def test_button_tag_with_block_and_text_passes
-          source = <<~RUBY
-            class TestView < Phlex::HTML
-              def view_template
-                button_tag(class: "button-icon") do
                   inline_svg("icon.svg")
                   span { t(".label") }
                 end
@@ -422,7 +300,7 @@ module A11y
         end
 
         def run_linter(source, filename: "test_view.rb")
-          PhlexRunner.new([MissingAccessibleName]).run(source, filename:)
+          PhlexRunner.new([LinkToMissingAccessibleName]).run(source, filename:)
         end
       end
     end